requirements.txt
numpy==1.21.3
psycopg2-binary==2.9.1
validate-docbr==1.10.0
